requirement.txt
Django==4.1
Pillow==9.2.0
django-ckeditor==6.5.0
